Implent the calculator UI to use class
